Test that we can make a sample molecule for every atom type.
This unit test was inspired by efforts to debug something in some
changes to the database, which caused a segfault when running 
the database Tests, but was in fact caused by an undetected bug in RMG. This should detect similar bugs.
The idea is that for every atomType, such as 'Cds', the
method Group().make_sample_molecule() can create a molecule
from the group definition, eg.
  1  Cd  ux
should make the molecule
  1 C u0 p0 c0 {2,D} {3,S} {4,S}
  2 C u0 p0 c0 {1,D} {5,S} {6,S}
  3 H u0 p0 c0 {1,S}
  4 H u0 p0 c0 {1,S}
  5 H u0 p0 c0 {2,S}
  6 H u0 p0 c0 {2,S}

Many work, but quite a few fail.
Updated to RMG3/Python3.