Unit test that Surface_Adsorption_vdW doesn't adsorb radicals.
Trying to forbid them using the forbidden structures doesn't 
seem to be working (a bug in is_subgraph_isomorphic?) but
making the top level node have multiplicity [1,3] excludes
anything with multiplicity 2 and seems to work on a 
generate_reactions level. Probably a better solution anyway, 
as it's inefficient to match the node, apply the recipe, generate
the products, and *then* discard them as forbidden.